Remove warnings: m8baby, m6/m8 warnings. 1. Most warnings of m8baby. 2. Warnings of each board. 3. Some PMU warnings by tao.zeng@amlogic.com 4. Some Storage warnings by benjamin.zhao@amlogic.com

+++ common/cmd_dtbload.c
@@ -57,7 +57,7 @@ int do_dtbload(cmd_tbl_t *cmdtp, int flag, int argc, char * const argv[])
                 return -1;
             }
             else
-                printf("load dtb file to memory 0x%x\n",dest_addr);
+                printf("load dtb file to memory 0x%x\n",(unsigned int)dest_addr);
         }
     }
 #endif
